Move setting window global to CTW provider
