+++ src/components/core/ctw-provider.tsx
@@ -7,12 +7,14 @@ import {
   ReactNode,
   useCallback,
   useContext,
+  useEffect,
   useMemo,
   useRef,
   useState,
 } from "react";
 import { CTWState, CTWStateContext, CTWToken } from "./ctw-context";
 import "./main.scss";
+import { version } from "../../../package.json";
 
 export type Env = "dev" | "sandbox" | "production";
 
@@ -28,9 +30,23 @@ type CTWProviderProps = {
   headers?: HeadersInit;
 } & (AuthTokenSpecified | AuthTokenURLSpecified);
 
+declare global {
+  interface Window {
+    CTWComponentLibrary: {
+      version: string;
+    };
+  }
+}
+
 function CTWProvider({ theme, children, ...ctwState }: CTWProviderProps) {
   const [token, setToken] = useState<CTWToken>();
   const ctwProviderRef = useRef<HTMLDivElement>(null);
+
+  useEffect(() => {
+    window.CTWComponentLibrary = {
+      version,
+    };
+  }, []);
 
   const handleAuth = useCallback(async () => {
     if (ctwState.authToken) {
